Allow resuming simulation by regularly saving to file
- Adds autosave_prefix and autosave_dt config fields
	By default:
		emu_mps_save_<uuid>.dat
		every 10 minute
- Accidentally fix the various config entries and
carry the whole MPSConfig along. It was part of my
research for optimal implementation and takes too
much time to separate as a independent MR.
- Some fixes to the unpickled impl are necessary,
like autosave_file (to save to same existing file)
and the logger which gets messed up by the pickling.
- Fix make_H modifying its interaction_matrix argument.
That was also part of the process but in fact now
unrelated.

---> How to resume a simulation from a file

import emu_mps
results = emu_mps.MPSBackend().resume("emu_mps_save_xxx.dat")