Fix baths not being updated after quantum jump
This was detected thanks to an assert regarding baths sizes.
Post-jump orthogonalization of the state would apply qr
decompositions and change state factor sizes (in the presence of very low
precision), without right baths being updated.

Originally an issue posted on Slack:
https://pasqalworkspace.slack.com/archives/C07MUV5K7EU/p1742474264332599

The right baths are not recomputed when root-finding, because the hami
ltonian doesn't change between iterations (within one timestep).
However, a jump implies applying an operator to the state: the right
baths need to be updated.