Make deallocate_tensor safer and work with views
deallocate_tensor would not work with views.
For that the _base tensor of a view needs to have
its storage set as well.

I added an exception when `deallocate_tensor` could not free
the memory because some other thing was pointing to
the data, using _use_count().

This allowed finding cases where the data was not properly allocated
so this change further reduces max memory usage in emu-mps.